[ Web ] Feat Changed location, add_schedule
